backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.1.10
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
flake8==4.0.1
install==1.3.5
mccabe==0.6.1
numpy==1.22.1
pandas==1.4.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
django-dbml==0.5.0
drf_spectacular==0.21.1
gunicorn==20.1.0
Faker==12.2.0
django-q==1.3.9
google-cloud-storage==2.2.1
openpyxl==3.0.9

# display tracebacks correctly in parallel test runner
tblib==1.7.0

# code coverage
coverage==6.2
coveralls==3.3.1
jsmin==3.0.0
django-extensions==3.1.5
XlsxWriter==3.0.3
django-mailjet==0.3.1

#debugger
debugpy==1.5.1
msgpack==1.0.4
requests-mock==1.11.0